Enemy Path Following
FollowPath and PathDefinition are the only scripts you need to make
something follow a path (doesn't have to be an enemy). WaterEleFacing
takes any game object (i used it specifically for the water elemental)
and compares its x position to the x position of the path checkpoint it
is headed towards and decides whether it should be facing left or right.